Changed ResidentsService to ResidentService, its path from residents to resident, getAll to list, getOne to retrieve; Updated imports to new path and class; Added docstrings to Resident related services, model, enums, pipes;~
