+++ web/src/app/shared/handlers/pipes/gender-type.pipe.ts
@@ -1,6 +1,19 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { GenderType } from '../../services/residents/residents.model';
+import { GenderType } from '../../services/resident/resident.model';
 
+/**
+ * Transforms an integer to a string representative of GenderType
+ * @see `GenderType`
+ *
+ * ### Pre-defined options
+ *
+ * | Option                  | Value            | String representative  |
+ * |-------------------------|------------------|------------------------|
+ * | GenderType.FEMALE       | 1                | Car                    |
+ * | GenderType.MALE         | 2                | Motorcycle             |
+ * | Else                    | Other than above | N/A                    |
+ * 
+ */
 @Pipe({
   name: 'genderType'
 })
