+++ web/src/app/shared/handlers/pipes/title-type.pipe.ts
@@ -1,6 +1,20 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { TitleType } from '../../services/residents/residents.model';
+import { TitleType } from '../../services/resident/resident.model';
 
+/**
+ * Transforms an integer to a string representative of TitleType
+ * @see `TitleType`
+ *
+ * ### Pre-defined options
+ *
+ * | Option             | Value            | String representative  |
+ * |--------------------|------------------|------------------------|
+ * | TitleType.MR       | 1                | Car                    |
+ * | TitleType.MRS      | 2                | Motorcycle             |
+ * | TitleType.MS       | 3                | Lorry                  |
+ * | Else               | Other than above | N/A                    |
+ * 
+ */
 @Pipe({
   name: 'titleType'
 })
