+++ web/src/app/shared/services/resident/resident.service.ts
@@ -4,14 +4,17 @@ import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 
-import { Resident } from './residents.model';
+import { Resident } from './resident.model';
 
 const BASE_URL = `${ environment.baseUrl }v1/residents/`
 
+/**
+ * A service for Resident related methods
+ */
 @Injectable({
   providedIn: 'root'
 })
-export class ResidentsService {
+export class ResidentService {
 
   // Data
   resident: Resident | undefined
@@ -21,7 +24,12 @@ export class ResidentsService {
     private http: HttpClient
   ) { }
 
-  getAll(): Observable<Resident[]> {
+  /**
+   * Returns a list of residents
+   *
+   * @returns List of residents
+   */
+  list(): Observable<Resident[]> {
     const urlTemp = `${ BASE_URL }`
     return this.http.get<Resident[]>(urlTemp).pipe(
       tap((res: Resident[]) => {
@@ -31,7 +39,14 @@ export class ResidentsService {
     )
   }
 
-  getOne(id: number): Observable<Resident> {
+  /**
+   * Returns a resident
+   * 
+   * @param id - resident ID
+   *
+   * @returns A resident
+   */
+  retrieve(id: number): Observable<Resident> {
     const urlTemp = `${ BASE_URL }${ id }/`
     return this.http.get<Resident>(urlTemp).pipe(
       tap((res: Resident) => {
@@ -41,6 +56,13 @@ export class ResidentsService {
     )
   }
 
+  /**
+   * Activate a resident
+   *
+   * @param id - resident ID
+   * 
+   * @returns Updated resident
+   */
   activate(id: number): Observable<Resident> {
     const urlTemp = `${ BASE_URL }${ id }/activate/`
     return this.http.get<Resident>(urlTemp).pipe(
@@ -51,6 +73,13 @@ export class ResidentsService {
     )
   }
 
+  /**
+   * Deactivate a resident
+   *
+   * @param id - resident ID
+   * 
+   * @returns Updated resident
+   */
   deactivate(id: number): Observable<Resident> {
     const urlTemp = `${ BASE_URL }${ id }/deactivate/`
     return this.http.get<Resident>(urlTemp).pipe(
@@ -61,6 +90,14 @@ export class ResidentsService {
     )
   }
 
+  /**
+   * Search residents through:
+   * - residentNo
+   * - name
+   * - email
+   * 
+   * @returns List of filtered residents
+   */
   search(searchText: string): Observable<Resident[]> {
     const urlTemp = `${ BASE_URL }?search=${ searchText }`
     return this.http.get<Resident[]>(urlTemp).pipe(
